fixed spelling of className
